+++ android/src/main/kotlin/com/lucasjosino/on_audio_query/types/ArtworkType.kt
@@ -8,7 +8,7 @@ fun checkArtworkType(sortType: Int): Uri {
     return when (sortType) {
         0 -> MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
         1 -> MediaStore.Audio.Albums.EXTERNAL_CONTENT_URI
-        else -> MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
+        else -> throw Exception("[checkArtworkType] value don't exist!")
     }
 }
 
@@ -16,6 +16,6 @@ fun checkArtworkFormat(format: Int) : Bitmap.CompressFormat {
     return when (format) {
         0 -> Bitmap.CompressFormat.JPEG
         1 -> Bitmap.CompressFormat.PNG
-        else -> Bitmap.CompressFormat.JPEG
+        else -> throw Exception("[checkArtworkFormat] value don't exist!")
     }
 }