+++ android/src/main/kotlin/com/lucasjosino/on_audio_query/types/AudiosFromType.kt
@@ -7,7 +7,11 @@ import android.provider.MediaStore
 fun checkAudiosFromType(sortType: Int): String {
     return when (sortType) {
         0 -> MediaStore.Audio.Media.ALBUM + "=?"
-        1 -> MediaStore.Audio.Media.ARTIST + "=?"
-        else -> MediaStore.Audio.Media.ALBUM + "=?"
+        1 -> MediaStore.Audio.Media.ALBUM_ID + "=?"
+        2 -> MediaStore.Audio.Media.ARTIST + "=?"
+        3 -> MediaStore.Audio.Media.ARTIST_ID + "=?"
+        4 -> MediaStore.Audio.Media.GENRE + "=?"
+        5 -> MediaStore.Audio.Media.GENRE_ID + "=?"
+        else -> throw Exception("[checkAudiosFromType] value don't exist!")
     }
 }