+++ lib/on_audio_query.dart
@@ -1,3 +1,16 @@
+/*
+=============
+Author: Lucas Josino
+Github: https://github.com/LucasPJS
+Website: https://lucasjosino.com/
+=============
+Plugin/Id: on_audio_query#0
+Homepage: https://github.com/LucasPJS/on_audio_query
+License: https://github.com/LucasPJS/on_audio_query/blob/main/LICENSE
+Copyright: © 2021, Lucas Josino. All rights reserved.
+=============
+*/
+
 library on_audio_query;
 
 //